Database Integration
I have completed the changes to make the DatabaseManager work within the
App.
On top of this all the changes have been made to the Activities and
Adapters to make the app work as a whole.
This has been debugged, however it requires testing, and a few functions
are missing. Specifically:
- Add a new store, however there is a spot to start the intent
- Notifications have been block-commented out of the DBMan as I don't
know what to do with the information to send and receive them from the
DBMan .