Stores and Filters
Hey guys. As per the conversation we had:
The stores page is there, and currently lists all the stores as well as
allows the searching for new stores. Awaiting Google Map functionality,
however instances of its use are isolated.
There are now filters in the ListAdapter for Store, Category and
"Today". The top element in the home screen was removed and replaced
with a relativelayout with some testing features to demonstrate.
As per the above the database has been updated.