Google Maps Integration. Untested
THIS IS A COPY OF THE STORES AND FILTERS COMMIT
I havnt built out of the latest commit because its bugged like crazy.
However I have the Google Maps started in the EditStore activity as well
as the StoresAdapter.
There have been a few other small changes throughout the project, namely
in the DatabaseManager as well as List and Stores models.

There is a non-fatal crash that happens when you load the GoogleMaps
component:
"Ensure that the "Google Maps Android API v2" is enabled." and it is
therefore untested, however the functionality is there and should work.